% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtilenew.R
\name{geom_tile_new}
\alias{geom_tile_new}
\title{Draw rectangles with wide white borders.}
\usage{
geom_tile_new(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created
by aes or aes_. If specified and inherit.aes = TRUE
(the default), is combined with the default mapping at
the top level of the plot. You only need to supply
mapping if there isn't a mapping defined for the plot.}

\item{data}{A data frame. If specified, overrides the default
data frame defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on
the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string,
or the result of a call to a position adjustment function.}

\item{...}{other arguments passed on to layer. There are three types of arguments you can use here.}

\item{na.rm}{f FALSE (the default), removes missing values
with a warning. If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in
the legends? NA, the default, includes if any aesthetics are
mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics,
rather than combining with them. This is most useful for helper
functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification}
}
\description{
\code{geom_tile} uses the center of the tile and its size (\code{x},
\code{y}, \code{width}, \code{height}).
}
\section{Aesthetics}{

}
\examples{
# To draw arbitrary rectangles with wide white borders, use geom_tile_new()
df <- data.frame(
  x = rep(c(2, 5, 7, 9, 12), 2),
  y = rep(c(1, 2), each = 5),
  z = factor(rep(1:5, each = 2)),
  w = rep(diff(c(0, 4, 6, 8, 10, 14)), 2)
)
ggplot(df, aes(x, y)) +
  geom_tile_new(aes(fill = z))

}

