% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomsqr.R
\docType{data}
\name{GeomDrawSquare}
\alias{GeomDrawSquare}
\title{@export}
\format{An object of class \code{GeomDrawSquare} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomDrawSquare
}
\description{
@export
}
\keyword{datasets}

