% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsfs7.R
\name{jsf_run_s7}
\alias{jsf_run_s7}
\title{Run JSF simulation and return S7 object}
\usage{
jsf_run_s7(x0 = c(10, 0), t_max = 10)
}
\arguments{
\item{x0}{Initial state vector}

\item{t_max}{Maximum simulation time}
}
\value{
JSFResult S7 object
}
\description{
Adds S7 encapsulation on top of core functionality
}
